Don't use the "Pop-up" video viewer thing during iframe embed.
First of all, if you're including the video as an iframe,
having some sort of javascript fake pop up window is probably
going to look really bad.

Second, it just simply doesn't work.

There's probably some additional issues in regards to how the
iframe feature works, but I think this is a sufficient fix
for now.

Bug: 56405
Change-Id: I70a54ef173f185e2cfd74f40982dc12f2af85c9f