Fix error reporting for ffmpeg2theora if don't recognize codec
Ffmpeg2theora still gives a return code of 0 even if transcoding
failed due to not recognizing the codec (e.g. VP9). Make it also
check that a non-zero byte file exists so that the error message
is reported properly

Previously the error message was reported as '1' (true cast to
string), as noted on bug 53863.

Change-Id: Id681b481c1a86d1d63528cf63c2a7dc5e430129b