Fix transcodes on files with long non-ascii names
Swift doesn't like it if the urlencoded version of the path
is > 1024 bytes, which includes files with names longer than
170 bytes that are non-ascii (Since each byte gets tripled to
percent encode, and the filename is repeated twice).

In these cases shorten the transcode name, much like how we
shorten thumbnail names for files with long names. Only do this
for files which would currently be broken, so we don't have to
worry about migrating existing transcodes.

Also remove backwards compatability in extension for versions
of MediaWiki not having File::getTranscodedUrl method.

Change-Id: I826af75ede734025352244608640a8b71929109a