Launch videos in VLC app on iOS if installed
iOS does not support ogg/webm. Direct people to install iOS app
instead, and direct them to view videos in that instead.

This is meant as a somewhat short term solution. After the
kaltura player update (whenever that happens), it may make sense
to replace this sort of thing with ogv.js, or perhaps keep this
vlc triggering code. Either way, in the near term this is a
relative simple fix that would help us reach 11.4% of our user base.

One minor issue: If you don't have the VLC app installed, you will
get a safari cannot open the page dialog box when we try to open
with vlc. I couldn't find anyway around it. I don't think its a big
deal, since in that case we cannot play the video anyways, so an
error is not out of place.

Bug: 61095
Change-Id: I6c3b5d8aadbd69577df3fe513f780f69708450d5