Override CSS link colour in pop up video
If using the pop-up player (from a really small sized thumbnail),
the link colours in the video are overriden to be super dark
grey on black. Which is just silly. This makes it hard to
read for the "VLCApp" player, and is just plain inconsistent.

This just restores the colours from the default MW css.

Change-Id: I2e45e45b60f118141a0a54276cfa397f54ea818f
Follow-up: 02a978c56b17