Don't try to use broken XiphQT-as-HTML5 in Safari
XiphQT components perhaps once worked in <video> and <audio> in
Safari but they don't as of Safari 8 on OS X 10.10.

Removed weird workaround that detected a registered ogg plugin
and claimed to support ogg natively in <video> and <audio>,
which resulted in just a blank screen (and even took priority
over ogv.js).

With this fixed, Safari will select ogv.js as expected.

Bug: T108251
Change-Id: I046814576efe04168e2eb75c9083caeee7bbe648