Add TMH modules to the upload page.
On the upload page, a fragment of wikicode is delivered as htmlblob by
iiprop uploadingwarning. This upload warning can contain thumbnails of
videos. Since there is no other way to currently do module loading for
this type of fragment, hardcode the page to always add the player's
loader modules. Depends on: I45e10f5

Bug: T51751
Change-Id: I4580b3a5c229d76a7cc61fde1a9383f80bc2a3f1