Don't list transcodes that were never initiated as failures.
Because we added a new transcode profile in I413a72be, almost every
single video/audio file is now listed on Special:TimedMediaHandler as
queued, because they were never initiated (relates to issue T104061),
and because the queued query was unspecific.
This makes the list less useful. I introduced a new state 'missing'
and narrowed 'queued', to stuff that was actually added to the
jobqueue.
Also recalculate totals, making sure failed, missing and currently
being transcoded are not part of the transcode total.

Bug: T132616
Bug: T112932
Change-Id: I9d96419b55c77924168ba14f2fe3a5ab3cac0698