Allow parser tests to be run in any order
Having a player ID which increments throughout script execution does not
allow individual parser tests to be run with --regex. Also, my pending
parser test engine refactor changes the order in which the tests are
run and so breaks tests. Instead, add a hook which resets the serial
at the start of each test.

Change-Id: I561344cee8e279050be37356fbfa8e97d3b21017