Optional BetaFeatures integration for video.js mode
If $wgTmhUseBetaFeatures is enabled, extends the BetaFeatures extension
with a switch which can turn on video.js mode per-user while the default
$wgTmhWebPlayer setting is still at 'mwembed'.

Beta-mode users get a parser-cache hash addition to ensure that caches
are not polluted.

Note that both mwembed and videojs RL modules are now registered together
and *loaded* conditionally. When we later retire mwembed mode we'll take
the old modules out.

Should have no visible effect when not enabled.

Bug: T100106
Change-Id: Ifae0fbc6883b679a115608191c432c0b8d0ca9ab