Fix length calculation for shortening transcode filenames
Follow-up to I826af75e to include the hash, which wasn't being
counted causing some files to be miscounted as not requiring
shortage. Also made the counting clearer on what it was doing
instead of having a random ' + 12' at the end. ;)

Should fix long Russian filenames that were _just_ the wrong
length to get stored.

Bug: T155248
Change-Id: I2a3765bd594f5c12b838f9c2d7593596c1dd3936