Cap .ogv transcode frame rate to 60 fps (1000 fps bug)
ffmpeg sometimes reports that WebM source videos are at 1000 fps
which ffmpeg2theora dutifully tries to process output for, creating
many duplicate frames.

This causes playback troubles, poorer compression, and much larger
encoding times, and is generally annoying as heck.

As a workaround, cap ffmpeg2theora's .ogv output to 60fps, which is
the highest actual frame rate we see in the wild. Along with the
'noUpscaling' option we already use this will not inflate files that
have a correct frame rate.

This will still create dupe frames on input that's really 30 fps or
whatever, but a LOT less. :)

Bug: T135671
Change-Id: Ie58aed3a3884fde9353c33233da400b8c0d35bad