More careful order & size check for video sources
Previously we always used the 'most preferred' codec if
available, even if the size was wildly inappropriate.
This caused slow playback of large VP9-encoded videos
even when a more suitable VP8 or Ogg was available.

Now discarding sources larger than the display size,
unless it's the only thing available.

Bug: T145925
Change-Id: I03ccd572ea386553df037e84ed896237c1f02e49