Enable WebM playback via ogv.js
WebM is now fairly stable in ogv.js, though current version spams console
a bit. Allow WebM to play back via the shim on Safari/IE/Edge.

This is prep for removing the Ogg Theora video transcodes from Wikimedia
in favor of WebM due to ongoing tooling issues.

Note that due to ordering of codecs, this will cause WebM versions to
play back preferably to the Ogg versions as soon as it goes out.

Bug: T172444
Change-Id: I68f22d7cb0cee4fe3f5742a742d45189b617fe07