Disable wasm, use asm.js codec modules for Safari/Edge
WebAssembly .wasm versions of the codec modules load faster, but
run about the same speed as the JS versions. Both are inscrutable. ;)

But having some trouble with the deployment and it should be tested
better first.

Also we should more thoroughly discuss implications of including
binary code modules for JS client-side code, and whether we should
be leaning on package managers for importing those.

Bug: T181022
Change-Id: I2ac297ca61fbf24c8c5b5428d3af5002b0925109