Remove Ogg Theora .ogv transcode output code paths
Since these are no longer used in production, and the conversions tools
are hard to maintain due to low general usage, going ahead and taking
out the Ogg Theora .ogv transcode output and ffmpeg2theora code paths.

Ogg Theora videos are still supported for playback and as transcode
sources, which will be converted to WebM or whatever is configured
as output.

The WebVideoTranscode:ENC_OGV_* constants are no longer defined,
so make sure they are removed from your LocalSettings.php

Bug: T181591
Change-Id: Ib98662ab8f8f9fe2ef70c1f7a79320285c3cda87