Order types for optimized video.js playback
Reorder sources for audio files so non-native playback is last,
and native playback is preferred, otherwise in same order as before.

Result on current output is:
* Safari, IE, Edge-without-media-codec-pack play MP3
* Firefox, Chrome, Edge-with-media-codec-pack play Ogg

Hypothetically something with either native MP3 nor Ogg would try
to play the Ogg via ogv.js. :)

Note the ogv.js library is still loaded, but the codecs aren't
initialized if not used.

Change-Id: I6c3a6e4e5ad01a33668d77135a19acc433c54885