Enable WebAssembly (wasm) version of ogv.js
The WebAssembly version of ogv.js is noticeably faster on current
Safari and Edge versions than the asm.js build; use it when
support is available, through ogv.js's own feature detection.

Initial problem on earlier deployments should now be fixed (T181022).

Secondary problem with iOS 11.2.2-6 is worked around upstream
in ogv.js with a bug detection check.

Note that it would still be nice to add to the best practices
for wasm usage on T181451, but we are following them roughly
as they're listed so far. :)

Bug: T179108
Change-Id: I3ac8531720feaaab55c1b7d14dece507f9ed3b43