Add $wgFFmpegVP9RowMT option to enable -row-mt for VP9 transcodes
Adds configuration option $wgFFmpegVP9RowMT to enable macroblock row
multithreading for VP9 transcodes (-row-mt 1).

This allows increasing $wgFFmpegThreads to a larger number for VP9
transcodes; without it encoding will be limited to 4 threads for HD,
or less for low resolution. With it enabled, larger thread counts like
8, 16 or 32 will work at all resolutions.

This requires libvpx 1.7 and a matching build of ffmpeg 3.3 or higher
or it will fail with an unrecognized option error.

Note that the stock ffmpeg in Debian stable doesn't yet include this
option, but the one in testing and unstable does.

Bug: T190334
Change-Id: Ic99c3b0a2aab9b2f56529d5b5dc8814f0475a647