Fixes for parsoid-rendered previews in NWE / 2017 wikitext editor
In MwEmbedPlayer mode, players were getting resized weirdly and
then disappearing durig preview in the new VE-backed wikitext editor.
This is due to the old MwEmbedPlayer making assumptions about where
sizes are set and such and having weeeeird markup internally.

(Things work cleaner under the new videojs-based mode which needs
some more work to reach production.)

Fix is to check before the transformation that things are correctly
listing the MwEmbed-specific bits, and if not it adds them in a way
that seems to work on the parsoid-borne markup.

Note that the transforms only apply if the page was loaded from
a view with video on it; that's a separate issue which can complicate
testing.

Bug: T198266
Bug: T198103
Change-Id: I005b183f488588cfeb853635cbd939ab12cd9fd7