Workaround for job queue reporting 0 length (T200813)
Job queue in production reports 0 length when queried, which was
breaking the throttle for requeueTranscodes.php.

Switch the queue size check to use the transcode table, which
doesn't distinguish between the two priority subqueues but should
be good enough for now.

This doesn't solve bug T200813 but works around it for TimedMediaHandler.

Change-Id: I724e640fd33a3090531c50dcd5b852abac9febf4