NewMwEmbedResourceManager: Remove indirection around module directories
This is now only used within TMH itself. This indirection was fragile and
overly complex, primarily because it (once) had to support injection of
modules from other extensions, but these have all been merged into TMH now,
which makes all this indirection obsolete.

The chain starts at NewMwEmbedResourceManager::register().
This previously accepted a partial directory path relative to $IP.
This had the downside of having to guess the filename of the entry
point within it, and required hardcoding the extensions' own
location (thus being incompatible with various dev setups that
mount /extensions elsewhere).

It now takes the absolute file path to the entry point.

The manual checks and exceptions were removed in favour of
'require', which does this natively.

Bug: T202017
Change-Id: I9818096dd2dca0d4a946a4f948d19940285d3fa7