NewMwEmbedResourceManager: Remove non-standard 'messageDir' option
Instead, set the standard ResourceLoaderFileModule 'messages' key
directly in the array, as usual.

This keeps for now the fact that the list of message key is
computed from en.json, but makes this more explicit.

Also remove tolerance for en.json not existing or not being readable.
If that is the case, an error should be produced. In addition to the
Exception thrown, this may cause file_get_contents to emit a warning,
which is fine.

Bug: T202017
Change-Id: I057b2f3be02c7d947536b37212ce9a0d94635ab4