Guard notice in Ogg Vorbis comments with no "=" char
Someone appears to have been trying to upload Ogg Vorbis files with
damaged comment fields, triggering a notice in the processing.

Use the ?? operator to guard against the missing second element in
the explode()d array for this case. It was already turning into
an empty string thanks to a cast, but this avoids the notice.

Bug: T230158
Change-Id: Ic6a2914c73c6686ad44a29fdec763306a623849f