Change default for $wgFFmpegLocation to '/usr/bin/ffmpeg'
Previously $wgFFmpegLocation defaulted to '/usr/bin/avconv' which
pointed to the old avconv fork of ffmpeg that was available some
years back as default in Ubuntu.

Debian and Ubuntu now use mainline ffmpeg, so updating the defaults
to make it easier to set up a TMH configuration on typical servers.

Note that old avconv may or may not still work, it's not even tested
anymore, so I've removed references to it in config docs.

Change-Id: I0194ef369fff385c766fea665ddb727f79cd0e00