Lazy-load JS payload & use OOUI dialog in video.js playback mode
Retools the video.js player mode to defer loading of video.js libraries,
and instead wrap the native players with loader links that will pop up
an OOUI dialog to run the video.js player in.

This returns us to similar behavior to use of the old popup players but
with cleaner base HTML that works when no JS is available, and keeps the
JS and CSS payloads relatively light on load when no player is used.

The dialog will be expanded in future with share/embed, download, etc
workflows, bringing full feature parity with the old jquery UI dialog.

The new OOUI dialog works on desktop and mobile, and is responsive in
terms of auto-filling the window/screen on small devices. A progress
bar is shown during initial module load to help perceived performance
on slow networks.

Audio players currently use the native default controls instead of
video.js transformations, to avoid adding a popup there. This will
change in future commits.

One known regression: audio captions are temporarily disabled until
the audio handling is set up to route to the dialog.

Future work to follow up on this will include:
* test failover behavior in legacy browsers that won't play
* handle audio with a better placeholder and use the dialog
* add share/embed and info in the dialog

Bug: T228339
Change-Id: Ic3bc46046d52de9d023a305867e815833f10cc24