Reducing size of video.js payload
1) using core distribution without the http-streaming plugin
(we will use that in the future but don't need it now)

This reduces payload size by about half, from 800k to <400k.

2) using pre-minified version which is more aggressive than
our minification

Further reduces payload size to 270k or so.

This is as reported by mw.inspect() so doesn't count gzip
compression.

Also upgraded to video.js 7.7 which includes our patch for
lazy-loading subtitle tracks. Config is updated to enable
the lazy-load option (the default behavior remains the same).

Bug: T228339
Change-Id: I4e9202eeeea667bf8f41e1e154352f3af82bd4dd