Transform audio players in video.js mode again
Provides a lazy-load placeholder for audio items that should work
with existing styling, and allows for captions on audio again.

Since the captions run directly through video.js now, the plugin
for showing them in a separate toast div is removed.

An icon is used as a poster image for audio so the dialog looks
more natural, but still has the space for captions if available.

Using emulated text tracks for now as that gets Safari native
audio captions working.

Also added a fix for ?embedplayer=yes mode to force the transform
to replace the placeholder inline when played, instead of going
into a popup. This is used on the iframe embed mode only for now.

Improvements left for later:

On audio, subtitles sometimes display partially under the control
bar for the player. (They move up if the mouse cursor is active,
so it seems to be making allowance for the video behavior!)
May need an upstream fix for that.

Safari native audio doesn't trigger on the first clickthrough;
it's losing the connection to the click gesture event from the
async action and doesn't seem to pick up the way ogv.js does.

Bug: T228339
Change-Id: If173471740fa1ef94640acbc3cbe7f30a3f29f2c