Switch iframeHook from ArticleFromTitle hook to MediaWikiPerformAction
The extension was hooking to ArticleFromTitle in order to
detect if the embedplayer=yes was set, output a player and
disable all the rest of the output. This was a misuse of the hook,
since the extension wasn't actually doing anything with the article,
but seems like it just needed an early-enough stage in the initialization
process to completely change the output.

MediaWikiPerformAction hook seems like a more appropriate place for
the extension to interviene: it's called late enough in the initialization
that all the nesessary stuff is fully initialized, all the redirects
are handled, but early enough to completely override what MediaWiki is doing
with the request. Seems appropriate according to the documentation as well.

Bug: T249162
Change-Id: I7ef6bd693af8b3cb950b1bd774ff88569b9cbb0f