Validate the configuration for transcoding
People often misconfigure or don't configure TMH's transcoding. This
is some basic validation of the configuration of the environment to
make it more clear that it is not correctly setup.

Questions
1. We don't have a ConfigurationException or something do we ?
2. This fails early. But the extension defaults now tend to fail the
entire MediaWiki, where before it did not. Do we change the defaults,
or do we force people to fix their installation ?
3. In case of the latter.. due to the early fail, this will now fail
any testcase, when TMH is installed and not configured.. How would we
override that ?

Bug: T260560
Change-Id: Ifdcc96434023300be40aadac04a949d690d4d20a