Use SkinTemplateNavigation::UniversalHook
This patch replaces the to-be deprecated SkinTemplateNavigationHook
(which does not run on special pages) with its more generic version
which runs on all pages.

The two hooks are directly interchangeable here because the extension
already does all sanity checks in ::isTimedMediaHandlerTitle() function
to ensure the TimedText tab appear only where it's safe and needed.

Bug: T255319
Change-Id: Ia650d5c373f8df4b7874a00f9fbde618d7c8df9c