Drop jquery.ui.touchpunch
The library  enables touch support for jquery ui. From the documentation
its a hack that duck punches jquery ui:
https://github.com/furf/jquery-ui-touch-punch
It has also not been updated since 2014.
We do not load this on mobile site as targets is not set.

For this reason I believe we should remove this library

Note because of the nature of this hack no other code needs to be removed.

The result of this change is that simulated events will no longer trigger
on touch devices which seems like a good thing to me given our need to move
away from jquery.ui and the fact we avoid loading it on the only supported
mobile skin.

The most noticeable result of this is that volume controls no longer work
with a swipe if a user is using Vector desktop on mobile and will now have to click it.
Seems acceptable.

Bug: T262947
Change-Id: I9c1534591a538078a0140429c6ffcecc8fa6f515