Remove stray decodeURIComponent call that could fail
There was a decodeURIComponent call on the input from video
attributes in the form data-mwtitle but no such encoding was
ever applied on the output, so this wasn't needed and can
fail.

Note the failure case in T268063 seemed to be Safari-specific,
which may indicate either browser-specific behavior in passing
through some characters, or that something in the ogv.js shim
support is somehow causing things to be handled differently.

I can't reproduce the bug on my local server, but simply
removing the stray call that doesn't belong there should help.

Bug: T268063
Change-Id: Iba6050cd481b67e92e7f9aeea65369eba4b51de2