build: Fix the video.js update bits
The script to update from npm had been left with some old
invocations to actually change the installed version, which
is not what we want I think -- we should be manually able
to control what version is used.

Removed 'npm update' invocations from 'grunt update-videojs'
and set the version to the currently used version – 7.7.4,
not 7.15.0.

Change-Id: I23ccbd058808bf50958df81c436ed157cf7a4040