Merge mw.PopUpMediaTransform.styles module into only user
The benefit of a separate style module is that it can be loaded
even when JavaScript is not enabled, but this module isn't, its
only use is as a dependency of the mw.PopUpMediaTransform.
Accordingly, there isn't a need to have a separate module, we
can just load the styles directly in mw.PopUpMediaTransform.

Bug: T280746
Change-Id: I9e421905ffdafceb04f85cc96c8092cc36e01756