Merge two dependencies into `mw.MwEmbedSupport` module
The `jquery.triggerQueueCallback` module was only used as a
dependency, and thus does not need to be separate.

The `jquery.mwEmbedUtil` module was used in two places, as a
dependency of `mw.MwEmbedSupport` and in
jquery.embedPlayer.js via `mw.loader.using()`, but that same
`mw.loader.using()` call also loads `mw.MwEmbedSupport`,
so if the two are merged the needed code will still be loaded.

Bug: T280746
Change-Id: I59b94dd498de1d2423b34cb42c739bc08d44f014