Merge mw.MwEmbedSupport.style module into only user
The benefit of a separate style module is that it can be loaded
even when JavaScript is not enabled, but this module isn't, its
only use is as a dependency of the mw.MwEmbedSupport.
Accordingly, there isn't a need to have a separate module, we
can just load the styles directly in mw.MwEmbedSupport.

Bug: T280746
Change-Id: I28b5b62de7d452c60bbba5532c9096b3d1a54a2f