Check if the native player already has the metadata loaded
There is this kludge in mwembed that checks if the metadata is loaded.
This is probably from before HTML5 video standardization was fully
completed. It forgets that sometimes this information can already be
available BEFORE this check runs. So check readyState and bail this
part of the flow if the metadata is already ready.

Bug: T295514
Change-Id: Ie2b1fea8b54f59dadfc45e78c0dcf9786a51afd8