Compensate subtitle display offset in audio player
In the audio player we force the control bar to always be visible.
The subtitle offset calculation from the bottom however depends on the
visibility of the controlbar. Because of this subtitles were showing
behind the controlbar.

If the player is inactive and the video is playing, push the subtitles
up by the controlbar height of 3em.

Bug: T294155
Change-Id: Id491166f2e006d4ddaba2e02425eba24d38fc747