Sidestep maint job bug with NS_TIMEDTEXT
A usage of NS_TIMEDTEXT constant in one of the special pages via
a maintenance script was causing intermittent failures, possibly
due to namespace canonical forms not having been loaded yet in
some scenario I can't reproduce locally.

Sidestepped it by using the config variable which is used to set
the var instead.

Bug: T287152
Change-Id: I7de6372865ac436ecf1f1ff962651c2c3194f0e3