VideoJS relayout fixes
There were some minor content relayouts as our JS loaded and replaced
the original video elements.

Mostly this was because of the border being removed icw responsive
sizing, but also because some of the native controls use box-sizing
border-box in some browsers. We now replace the border with a
1px margin outside the boxmodel.

Additionally I discovered an issue with Chrome not respecting A/R of
video elements when you insert them late into the DOM. Not entirely
sure if this is because of the AR they set from the useragent side,
which is different from most browsers in that it uses 'auto width /
height'. It might be that intrisic sizes of replaced content work
differently from my understanding, or that there is simply an error in
their implementation, where it only sets the A/R after the poster has
loaded in. By overriding the A/R to a preferred A/R using inline
styles in our MediaElement class, we can compensate for this.

Change-Id: Iee234df69bc57220a25698f9b702803a5c59c914