Consistently use MainConfig service in hooks
And remove the unused 'timedmediahandler' config entry, which per
T249564 was part of an early experiment that gained no traction and
is currently incomplete and doesn't serve any purpose yet since it's
essentially a no-op that effectively just uses MainConfig still, but
through a second instance of GlobalVarConfig.

The motivation for this is to simplify the code before refactoring
TMH hooks refactoring around ParserCache (T230505).

Bug: T249564
Bug: T230505
Change-Id: Id58366e5dbe569fca60f59eed3920aac28482adb