Size the playbutton on videos smaller than 60px
The platbutton is sized to be min. of 54px. Unfortunately this causes
it to overflow on certain landscape videos of small width, esp.
in the upload history. Apply some sizing to allow for smaller videos.

This still breaks down when the size drops to around 30px, as the
included SVG has a 20x20 inherent size, but should be good enough.

Change-Id: Ie9909e8c0bf2597086f9137000012eb888629e0e