Disable poster and bigplaybutton in dialog
When we open the dialog we autoplay and there is no need for the
big play button, so explicitly disable it (and remove CSS which did
this before).

In addition remove the poster from the video we open in
the dialog. This caused potentially many flashes of content, which
is distracting. We had 'dialog', <video> poster, videojs poster,
videojs start of playback. This is not helped by the fact that
for TMH we generally create a poster slightly into the video, instead
of start of video (because frame 0 is often not very interesting).

Bug: T300400
Change-Id: I47cc43e7dbf4936fe7c949adfca73a6767bac14b