Workaround for audio with ringer off on iOS
Had to copy in the fix from ogv.js itself because we create the
audio context proactively to allow lazy-loading. This is simply
dropping a short silent <audio> in and playing it before creating
the audio context, which fixes the output channel for the mixer.

Note that this fix should be independent of the ogv.js update
itself, and can be cherry-picked separately in a pinch.

Bug: T301741
Change-Id: I93c056190610f625d8fe609f1185b22717567f19