Mark api timedtext 'trackformat' param as required
This avoids throwing an exception for invalid requests, which should
not be produced by TimedMediaHandler old or new versions but can be
constructed manually.

If there are still these broken URLs being used and they're not
just people experimenting with the API, please check for anything
that might be generating them wrong.

Bug: T306527
Change-Id: I73670ce02cfe30d635a970fe00b8d818b04160b4