Mark TMH modules as requiring ES6
We've had support using ES6 and crashing on IE11 for well over half a
year and there hasn't been a single complaint. As upstream has
switched to ES6 as well, we might as well flip this switch officially
Will also make it possible to cleanup the module code with ES6
conventions.

Bug: T307075
Bug: T306711
Change-Id: I072dfc74905458f7683cd604387990ddeb09f986