Fix for ogv.js enabling on iOS with MPEG source videos
Safari reports canPlay('video/mpeg') === 'maybe', then rejects
MPEG-1 and MPEG-2 source files based on their codecs. However
we've already used the 'maybe' coerced to a boolean value to
decide not to load ogv.js: we can't load any of the VP9 WebMs
but it tries to load the MPEG, fails, and has no failover.

Tweaking the check in isMediaNativelySupported to only accept
true (very old browsers) or 'probably' from canPlayType is more
accurate for this case.

Change-Id: Iba7466f20e8b67d571b7203894fa0f8015aac54f