Swap the media element with a span
In 2bfadd2, we used the span that wraps the media element as the
mw-tmh-player span.  That was problematic because sizing is placed on
the span explicitly and the styling in core doesn't expect it.  Here we
swap the element itself, as is done on the legacy path, and adjust the
styling in core to match spans as well so that it will be styled
appropriately.

Bug: T304343
Bug: T314097
Depends-On: I7bcde25aa3b319ff93f3ef40361682824d171ea8
Change-Id: I55bbd4f6b72f233a9481e9abd68c2863cf24277b