Revert "Video transcode config cleanup"
This reverts commit 21a48ef249a0afcbff44aa44794cd226436d8304.

Make the soon-to-be-released REL1_39 consistent with current master, not landing a feature that has been reverted on master.

Reason for revert: Caused all transcodes in WMF production to fail
as it depends on a flag in a newer version of ffmpeg than we have,
and ffmpeg fails noisily rather than continues when it is given a
flag it doesn't recognise.

Bug: T312152
Bug: T317069
Change-Id: I918a19de32903876da471f41268e8fdae0e4f7f5
(cherry picked from commit f09b04c7a584a146276473c2b7762fe48b6cc017)