Fix the download url
Need to use download attribute and/or tell the swift backend or other
server to download, otherwise it loads the web player in the browser if
the browser supports that type

Note that the download attribute only works same-origin, while the
?download query string parameter is not standardized and will work
on sites configured as Wikimedia sites are, but not on every site.

If the track links are off-origin, and the query-string has no effect,
then behavior will transparently revert to the previous behavior where
playable formats may play inline.

Change-Id: Ic89cbb477123e607630976b9af77c730f9238650