JSDoc for TMH modules
Documentation for this is hard and messy with all these globals
This should do for now.

Also made a few minor code changes to videojs resolution switcher
One function was returning a value in 'some' conditions, and there was
functionality for a callback which we never used,
so better off removing it.

Bug: T338458
Change-Id: Ic6b49023f8261892e862fd376eca9c32f15819a2