Build: fix jsdoc warning
Fix `npm run-script doc` which fails due to the following warnings:

 WARNING: Unknown link Object.<string, Array.<SourceObject>> in Global
 WARNING: Unknown link Object.<string, Array.<SourceObject>> in Global
 WARNING: Unknown link Object.<string, Array.<SourceObject>> in Global

We could potentially add an entry for `Object` to the linkMap but I find
it nicer to simply skip using `Object` in the affected type definition.

The rendering in `docs/js/VideoJsResolutionSwitcherPlugin.html` for the
`BuckectObject` type looks like:

  *BucketObject*

    Bucketized sources by label, res and type

  *Properties:*
    Name  | Type   | Description
    ------+--------+------------------------------------------------
    label | Object | Dictionary of labels to sources with that label
    res   | Object | Dictionary of res to sources with that res
    type  | Object | Dictionary of type to sources with that type

The sole different is the properties are marked as type `Object` (which
maybe should be their own type definitions?).

Bug: T338458
Change-Id: Ida2191140701b3ddcd2535a26cd877a81a52a391