Fix audio player being too wide when !parserEnableLegacyMediaDOM
Restores a bit configuration that was removed in 30b18d1.

When wildcard selectors were removed in
I70c61493fe492445702f036e5b24ef87fc3bdf43 the placeholder span where the
width is placed for audio elements no longer had the styling applied to
it.  Here we transfer the mw-file-element class from the audio element
when we wrap to transfer the styles.

Bug: T304343
Change-Id: I3a5ee7fde8bdc146c828cf89a02c5c98049e629c