Attempting to work around newly trapping SecurityCheck-XSS
The phan SecurityCheck-XSS check triggers in the iframe output mode
in TMH now, this seems to be an upstream/core/tools changs and not
anything new in TMH.

Not quite sure why it's triggering though, as it seems to complain
about output from OutputPage::getHeadLinksArray which itself is
made up of output from Html::element and should have proper taint
markings for escaped output.

Trying to mark the array of HTML chunks as properly escaped.

Change-Id: Id00a64a8156f13ae2a5b846b1341147a87fd504b