Don't try to lock to serialize m3u8 file writes
This caused problems in practice because there's a lot of weird
commit/transaction scenarios that can wrap these calls, and
sometimes it didn't allow for locking safely.

It might be wise to serialize updates cleanly through the job
queue or something in a way that doesn't force web jobs to
wait on them, but simply letting the last one "win" is good
enough for now and has a good failure mode (updating state
can recreate the file later in proper condition).

Additionally, exceptions are converted to a Status object
being passed back, which can be queried for internal debugging
but which won't interfere with production activity.

In the future, a clean debugging log for these failures may
help.

Bug: T348689
Bug: T348667
Bug: T348375
Bug: T348753
Change-Id: Idcfd6136eb835855a6ce7c2a6f5ad9fb5c37a02b