Don't change transcode rows during read operations
There was some old lazy-loading attempts to update the transcode
table entries with timeout error messages during read operations
which was causing trouble with our multi-dc system.

These db updates are not necessary, as the requeue scripts are
capable of deciding on their own that an old entry should be re-run.

Bug: T152851
Change-Id: I32a1f7787df2d60a3035dd866a31653184b14f6a