Always load transcode state from db when opting in to primary db
and force data file reads to READ_LATEST in WebVideoTranscodeJob

These are for write operations, or reads before writes, and
need to be up to date. May be involved in some wonky state
checks with replication (may still need work).

This _should_ help with unreplicated file metadata breaking some
transcodes when they start in the production environment (seen
in error returns as attempting to resize to '0x0' dimensions
or fixed-size mjpeg with ':' aspect ratio)

Fixed a typo in an ApiTranscodeReset method name; it's marked
public but used internally only so should be safe.

Bug: T200939

Change-Id: Ib86645f50f752e46dd548565460bd1e28c8c86c3