Preload auto when playing in dialog
When opening a player from the dialog, we know that the user is
committing to playing the file. Thus load videojs with the preload
option set to auto. This signals to the UA that it is allowed to start
downloading the complete file if it wants to.

This might speed up playback during the initial playback phase, avoiding
buffering pauses etc.

Change-Id: I303a49f221c093bc33daf7d4d1ea363e0ab39fc3