iOS 10-15 compat fix for HLS .m3u8 playlists
In iOS 16 the system HLS player starts using actual availability of
codec support to filter the specified codecs in the playlist; in
iOS 15 and earlier it appears to use a hardcoded list and would
filter out old but working codecs like 'jpeg' or 'mp4v'.

Simply skipping the CODECS specifier for the back-compat combo of
'jpeg' or 'jpeg,mp4a.6b' gets this working in iOS 15 and doesn't
regress HLS playback on iOS 16/17, where the higher resolutions
and more efficient bitrates of VP9 always take priority on newer
supporting devices.

Should fix media playback in the app on iOS 15, which was previously
broken for the new HLS videos even though it could play back the
MJPEG/MP3 fallback codecs.

Also works on mobile web as far back as iOS 10, the minimum required
to support fragmented MP4 track files in HLS. Note that the same
technique should work with h.263 or MPEG-4 Visual codecs, which give
much better quality and appear to be out of patent. (If we add h.264
in the future, it can be correctly labeled and will work back to
iOS 10 as well.)

Added rebuildPlaylists.php maintenance script to force recreation of
.m3u8 multivariant playlists when we change format or codecs. This
will need to be run after the fix is pushed to update existing
files. New uploads and newly generated transcodes will have their
playlists generated correctly once the fix is live.

Change-Id: Ie287619a2ac922c433b3444bba5fb8806397afd6
Bug: T354502