Fix "Unknown status" transcodes after move or reupload
I noticed that whenever this situation occured, we had entries, but
these entries always had a null addjob timestamp.

The only way that can really happen is if a page purge triggered
an update of the transcode table. My theory is that with the multiple
jobrunners in WMF, a page purge job (due to the move) is triggered
before all the jobs are added by updateJobQueue.

The easiest solution is to upsert, instead of doing an insert.
Then if there was an insert or updated row, we create the job.

Additionally, ensure that we start work, if the addjob time is newer
than the startwork time already currently set.

Bug: T113191
Change-Id: I647bed1540b5b45be8f73cd9b8b3cd7886e5caf4